+++ src/m365/spo/commands/group/group-user-add.spec.ts
@@ -12,6 +12,7 @@ const command: Command = require('./group-user-add');
 describe(commands.GROUP_USER_ADD, () => {
   let log: string[];
   let logger: Logger;
+  let loggerLogSpy: sinon.SinonSpy;
 
   const JSONSingleUser =
   {
@@ -40,17 +41,19 @@ describe(commands.GROUP_USER_ADD, () => {
     ]
   };
 
-  const JSONErrorResponseInvalidGroup =
+  const JSONGroupNotFound =
   {
-    ErrorMessage: "Group cannot be found.",
-    IconUrl: null,
-    InvitedUsers: null,
-    Name: null,
-    PermissionsPageRelativeUrl: null,
-    StatusCode: -9,
-    UniquelyPermissionedUsers: null,
-    Url: null,
-    UsersAddedToGroup: null
+    status: 404,
+    statusText: "Not Found",
+    error: {
+      "odata.error": {
+        code: "-2146232832, Microsoft.SharePoint.SPException",
+        message: {
+          lang: "en-US",
+          value: "Group cannot be found."
+        }
+      }
+    }
   };
 
   const JSONErrorResponseInvalidUsers =
@@ -100,11 +103,13 @@ describe(commands.GROUP_USER_ADD, () => {
         log.push(msg);
       }
     };
+    loggerLogSpy = sinon.spy(logger, 'log');
     // loggerLogToStderrSpy = sinon.spy(logger, 'logToStderr');
   });
 
   afterEach(() => {
     Utils.restore([
+      request.get,
       request.post,
       Cli.executeCommandWithOutput
     ]);
@@ -126,11 +131,74 @@ describe(commands.GROUP_USER_ADD, () => {
     assert.notStrictEqual(command.description, null);
   });
 
+  it('fails validation if both groupId and groupName options are passed', (done) => {
+    const actual = command.validate({
+      options: {
+        webUrl: "https://contoso.sharepoint.com/sites/SiteA",
+        groupId: 32,
+        groupName: "Contoso Site Owners",
+        userName: "Alex.Wilber@contoso.com"
+      }
+    });
+    assert.notStrictEqual(actual, true);
+    done();
+  });
+
+  it('fails validation if both groupId and groupName options are not passed', (done) => {
+    const actual = command.validate({
+      options: {
+        webUrl: "https://contoso.sharepoint.com/sites/SiteA",
+        userName: "Alex.Wilber@contoso.com"
+      }
+    });
+    assert.notStrictEqual(actual, true);
+    done();
+  });
+
+  it('fails validation if both userName and email options are passed', (done) => {
+    const actual = command.validate({
+      options: {
+        webUrl: "https://contoso.sharepoint.com/sites/SiteA",
+        groupId: 32,
+        email: "Alex.Wilber@contoso.com",
+        userName: "Alex.Wilber@contoso.com"
+      }
+    });
+    assert.notStrictEqual(actual, true);
+    done();
+  });
+
+  it('fails validation if both userName and email options are not passed', (done) => {
+    const actual = command.validate({
+      options: {
+        webUrl: "https://contoso.sharepoint.com/sites/SiteA",
+        groupId: 32
+      }
+    });
+    assert.notStrictEqual(actual, true);
+    done();
+  });
+
+  it('fails validation if webURL is Invalid', () => {
+    const actual = command.validate({ options: { webUrl: "InvalidWEBURL", groupId: 32, userName: "Alex.Wilber@contoso.com" } });
+    assert.notStrictEqual(actual, true);
+  });
+
+  it('fails validation if groupID is Invalid', () => {
+    const actual = command.validate({ options: { webUrl: "https://contoso.sharepoint.com/sites/SiteA", groupId: "NOGROUP", userName: "Alex.Wilber@contoso.com" } });
+    assert.notStrictEqual(actual, true);
+  });
+
+  it('passes validation if all the required options are specified', () => {
+    const actual = command.validate({ options: { webUrl: "https://contoso.sharepoint.com/sites/SiteA", groupId: 32, userName: "Alex.Wilber@contoso.com" } });
+    assert.strictEqual(actual, true);
+  });
+
   it('defines correct properties for the default output', () => {
     assert.deepStrictEqual(command.defaultProperties(), ['DisplayName', 'Email']);
   });
 
-  it('Adding user to a SharePoint Group', (done) => {
+  it('adds user to a SharePoint Group by groupId and userName', (done) => {
     sinon.stub(Cli, 'executeCommandWithOutput').callsFake(() => Promise.resolve({
       stdout: JSON.stringify(userInformation),
       stderr: ''
@@ -140,6 +208,15 @@ describe(commands.GROUP_USER_ADD, () => {
       if (opts.url === 'https://contoso.sharepoint.com/sites/SiteA/_api/SP.Web.ShareObject' &&
         opts.data) {
         return Promise.resolve(JSONSingleUser);
+      }
+
+      return Promise.reject(`Invalid request ${JSON.stringify(opts)}`);
+    });
+    sinon.stub(request, 'get').callsFake(opts => {
+      if (opts.url === `https://contoso.sharepoint.com/sites/SiteA/_api/web/sitegroups/GetById('32')`){
+        return Promise.resolve({
+          Id: 32
+        });
       }
 
       return Promise.reject(`Invalid request ${JSON.stringify(opts)}`);
@@ -153,6 +230,7 @@ describe(commands.GROUP_USER_ADD, () => {
       }
     }, (err?: any) => {
       try {
+        assert(loggerLogSpy.calledWith(JSONSingleUser.UsersAddedToGroup));
         assert.strictEqual(typeof err, 'undefined');
         done();
       }
@@ -162,11 +240,20 @@ describe(commands.GROUP_USER_ADD, () => {
     });
   });
 
-  it('Adding user to a SharePoint Group (DEBUG)', (done) => {
+  it('adds user to a SharePoint Group by groupName and email (DEBUG)', (done) => {
     sinon.stub(Cli, 'executeCommandWithOutput').callsFake(() => Promise.resolve({
       stdout: JSON.stringify(userInformation),
       stderr: ''
     }));
+
+    sinon.stub(request, 'get').callsFake((opts) => {
+      if ((opts.url as string).indexOf(`https://contoso.sharepoint.com/sites/SiteA/_api/web/sitegroups/GetByName(`) > -1) {
+        return Promise.resolve({
+          Id: 7
+        });
+      }
+      return Promise.reject('Invalid Request');
+    });
 
     sinon.stub(request, 'post').callsFake(opts => {
       if (opts.url === 'https://contoso.sharepoint.com/sites/SiteA/_api/SP.Web.ShareObject' &&
@@ -180,11 +267,12 @@ describe(commands.GROUP_USER_ADD, () => {
       options: {
         debug: true,
         webUrl: "https://contoso.sharepoint.com/sites/SiteA",
-        groupId: 32,
-        userName: "Alex.Wilber@contoso.com"
+        groupName: "Contoso Site Owners",
+        email: "Alex.Wilber@contoso.com"
       }
     }, (err?: any) => {
       try {
+        assert(loggerLogSpy.calledWith(JSONSingleUser.UsersAddedToGroup));
         assert.strictEqual(typeof err, 'undefined');
         done();
       }
@@ -194,19 +282,39 @@ describe(commands.GROUP_USER_ADD, () => {
     });
   });
 
-  it('Handles error when adding user to a SharePoint Group - Invalid Group', (done) => {
-    sinon.stub(Cli, 'executeCommandWithOutput').callsFake(() => Promise.resolve({
-      stdout: JSON.stringify(userInformation),
-      stderr: ''
-    }));
-
-    sinon.stub(request, 'post').callsFake(opts => {
-      if (opts.url === 'https://contoso.sharepoint.com/sites/SiteA/_api/SP.Web.ShareObject' &&
-        opts.data) {
-        return Promise.resolve(JSONErrorResponseInvalidGroup);
+  it('fails to get group when does not exists', (done) => {
+    sinon.stub(request, 'get').callsFake((opts) => {
+      if ((opts.url as string).indexOf(`https://contoso.sharepoint.com/sites/SiteA/_api/web/sitegroups/GetByName('`) > -1) {
+        return Promise.resolve({});
       }
+      return Promise.reject('The specified group not exist in the SharePoint site');
+    });
 
+    command.action(logger, {
+      options: {
+        debug: true,
+        webUrl: "https://contoso.sharepoint.com/sites/SiteA",
+        groupName: "Contoso Site Owners",
+        email: "Alex.Wilber@contoso.com"
+      }
+    }, (err?: any) => {
+      try {
+        assert.strictEqual(JSON.stringify(err), JSON.stringify(new CommandError(`The specified group does not exist in the SharePoint site`)));
+        done();
+      }
+      catch (e) {
+        done(e);
+      }
+    });
+  });
+
+  it('handles error when adding user to a SharePoint Group - Invalid Group', (done) => {
+    sinon.stub(request, 'get').callsFake(opts => {
+      if ((opts.url as string).indexOf(`https://contoso.sharepoint.com/sites/SiteA/_api/web/sitegroups/GetById('`) > -1 ) {
+        return Promise.reject(JSONGroupNotFound);
+      }
       return Promise.reject(`Invalid request ${JSON.stringify(opts)}`);
+
     });
     command.action(logger, {
       options: {
@@ -226,7 +334,7 @@ describe(commands.GROUP_USER_ADD, () => {
     });
   });
 
-  it('Handles error when adding user to a SharePoint Group - Username Does Not exist', (done) => {
+  it('handles error when adding user to a SharePoint Group - Username Does Not exist', (done) => {
     sinon.stub(Cli, 'executeCommandWithOutput').callsFake(() => Promise.reject({
       error: `Resource 'Alex.Wilber@invalidcontoso.onmicrosoft.com' does not exist or one of its queried reference-property objects are not present.`,
       stderr: `Resource 'Alex.Wilber@invalidcontoso.onmicrosoft.com' does not exist or one of its queried reference-property objects are not present. stderr`
@@ -236,6 +344,16 @@ describe(commands.GROUP_USER_ADD, () => {
       if (opts.url === 'https://contoso.sharepoint.com/sites/SiteA/_api/SP.Web.ShareObject' &&
         opts.data) {
         return Promise.resolve(JSONErrorResponseInvalidUsers);
+      }
+
+      return Promise.reject(`Invalid request ${JSON.stringify(opts)}`);
+    });
+
+    sinon.stub(request, 'get').callsFake(opts => {
+      if ((opts.url as string).indexOf(`https://contoso.sharepoint.com/sites/SiteA/_api/web/sitegroups/GetById('`) > -1) {
+        return Promise.resolve({
+          Id: 4
+        });
       }
 
       return Promise.reject(`Invalid request ${JSON.stringify(opts)}`);
@@ -257,21 +375,6 @@ describe(commands.GROUP_USER_ADD, () => {
         done(e);
       }
     });
-  });
-
-  it('fails validation if webURL is Invalid', () => {
-    const actual = command.validate({ options: { webUrl: "InvalidWEBURL", groupId: 32, userName: "Alex.Wilber@contoso.com" } });
-    assert.notStrictEqual(actual, true);
-  });
-
-  it('fails validation if groupID is Invalid', () => {
-    const actual = command.validate({ options: { webUrl: "https://contoso.sharepoint.com/sites/SiteA", groupId: "NOGROUP", userName: "Alex.Wilber@contoso.com" } });
-    assert.notStrictEqual(actual, true);
-  });
-
-  it('passes validation if all the required options are specified', () => {
-    const actual = command.validate({ options: { webUrl: "https://contoso.sharepoint.com/sites/SiteA", groupId: 32, userName: "Alex.Wilber@contoso.com" } });
-    assert.strictEqual(actual, true);
   });
 
   it('supports debug mode', () => {