+++ src/m365/teams/commands/tab/tab-get.spec.ts
@@ -307,7 +307,7 @@ describe(commands.TAB_GET, () => {
 
   it('fails to get team when team does not exists', (done) => {
     sinon.stub(request, 'get').callsFake((opts) => {
-      if ((opts.url as string).indexOf(`/me/joinedTeams?$filter=displayName eq '`) > -1) {
+      if ((opts.url as string).indexOf(`/v1.0/groups?$filter=displayName eq '`) > -1) {
         return Promise.resolve({ value: [] });
       }
       return Promise.reject('The specified team does not exist in the Microsoft Teams');
@@ -331,9 +331,60 @@ describe(commands.TAB_GET, () => {
     });
   });
 
+  it('fails when team name does not exist', (done) => {
+    sinon.stub(request, 'get').callsFake((opts) => {
+      if ((opts.url as string).indexOf(`/v1.0/groups?$filter=displayName eq '`) > -1) {
+        return Promise.resolve({
+          "@odata.context": "https://graph.microsoft.com/v1.0/$metadata#teams",
+          "@odata.count": 1,
+          "value": [
+            {
+              "id": "00000000-0000-0000-0000-000000000000",
+              "createdDateTime": null,
+              "displayName": "Team Name",
+              "description": "Team Description",
+              "internalId": null,
+              "classification": null,
+              "specialization": null,
+              "visibility": null,
+              "webUrl": null,
+              "isArchived": false,
+              "isMembershipLimitedToOwners": null,
+              "memberSettings": null,
+              "guestSettings": null,
+              "messagingSettings": null,
+              "funSettings": null,
+              "discoverySettings": null,
+              "resourceProvisioningOptions": []
+            }
+          ]
+        }
+        );
+      }
+      return Promise.reject('Invalid request');
+    });
+
+    command.action(logger, {
+      options: {
+        debug: true,
+        teamName: 'Team Name',
+        channelName: 'Channel Name',
+        tabName: 'Tab Name'
+      }
+    }, (err?: any) => {
+      try {
+        assert.strictEqual(JSON.stringify(err), JSON.stringify(new CommandError(`The specified team does not exist in the Microsoft Teams`)));
+        done();
+      }
+      catch (e) {
+        done(e);
+      }
+    });
+  });
+
   it('fails when multiple teams with same name exists', (done) => {
     sinon.stub(request, 'get').callsFake((opts) => {
-      if ((opts.url as string).indexOf(`/me/joinedTeams?$filter=displayName eq '`) > -1) {
+      if ((opts.url as string).indexOf(`/v1.0/groups?$filter=displayName eq '`) > -1) {
         return Promise.resolve({
           "@odata.context": "https://graph.microsoft.com/v1.0/$metadata#teams",
           "@odata.count": 2,
@@ -354,7 +405,8 @@ describe(commands.TAB_GET, () => {
               "guestSettings": null,
               "messagingSettings": null,
               "funSettings": null,
-              "discoverySettings": null
+              "discoverySettings": null,
+              "resourceProvisioningOptions": ["Team"]
             },
             {
               "id": "00000000-0000-0000-0000-000000000000",
@@ -372,7 +424,8 @@ describe(commands.TAB_GET, () => {
               "guestSettings": null,
               "messagingSettings": null,
               "funSettings": null,
-              "discoverySettings": null
+              "discoverySettings": null,
+              "resourceProvisioningOptions": ["Team"]
             }
           ]
         }
@@ -401,7 +454,7 @@ describe(commands.TAB_GET, () => {
 
   it('should get a Microsoft Teams Tab by Team name', (done) => {
     sinon.stub(request, 'get').callsFake((opts) => {
-      if ((opts.url as string).indexOf(`/me/joinedTeams?$filter=displayName eq '`) > -1) {
+      if ((opts.url as string).indexOf(`/v1.0/groups?$filter=displayName eq '`) > -1) {
         return Promise.resolve({
           "value": [
             {
@@ -420,7 +473,8 @@ describe(commands.TAB_GET, () => {
               "guestSettings": null,
               "messagingSettings": null,
               "funSettings": null,
-              "discoverySettings": null
+              "discoverySettings": null,
+              "resourceProvisioningOptions": ["Team"]
             }
           ]
         });