+++ src/settingsNames.ts
@@ -2,7 +2,12 @@ const settingsNames = {
   errorOutput: 'errorOutput',
   output: 'output',
   printErrorsAsPlainText: 'printErrorsAsPlainText',
-  showHelpOnFailure: 'showHelpOnFailure'
+  showHelpOnFailure: 'showHelpOnFailure',
+  csvHeader: 'csvHeader',
+  csvEscape: 'csvEscape',
+  csvQuote: 'csvQuote',
+  csvQuoted: 'csvQuoted',
+  csvQuotedEmpty: 'csvQuotedEmpty'
 };
 
 export { settingsNames };